Move all public nested classes out into own files
