+++ Terminal.Gui/ConsoleDrivers/FakeDriver/FakeDriver.cs
@@ -3,7 +3,6 @@
 //
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Threading;
@@ -13,31 +12,16 @@ using System.Text;
 using Console = Terminal.Gui.FakeConsole;
 
 namespace Terminal.Gui {
+
 	/// <summary>
 	/// Implements a mock ConsoleDriver for unit testing
 	/// </summary>
 	public class FakeDriver : ConsoleDriver {
 #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
 
-		public class Behaviors {
+		
 
-			public bool UseFakeClipboard { get; internal set; }
-			public bool FakeClipboardAlwaysThrowsNotSupportedException { get; internal set; }
-			public bool FakeClipboardIsSupportedAlwaysFalse { get; internal set; }
-
-			public Behaviors (bool useFakeClipboard = false, bool fakeClipboardAlwaysThrowsNotSupportedException = false, bool fakeClipboardIsSupportedAlwaysTrue = false)
-			{
-				UseFakeClipboard = useFakeClipboard;
-				FakeClipboardAlwaysThrowsNotSupportedException = fakeClipboardAlwaysThrowsNotSupportedException;
-				FakeClipboardIsSupportedAlwaysFalse = fakeClipboardIsSupportedAlwaysTrue;
-
-				// double check usage is correct
-				Debug.Assert (useFakeClipboard == false && fakeClipboardAlwaysThrowsNotSupportedException == false);
-				Debug.Assert (useFakeClipboard == false && fakeClipboardIsSupportedAlwaysTrue == false);
-			}
-		}
-
-		public static FakeDriver.Behaviors FakeBehaviors = new Behaviors ();
+		public static FakeDriverBehaviors FakeBehaviors = new FakeDriverBehaviors ();
 
 		int cols, rows, left, top;
 		public override int Cols => cols;
