+++ Terminal.Gui/Views/GraphView/Series.cs
@@ -72,7 +72,7 @@ namespace Terminal.Gui {
 
 		/// <summary>
 		/// The number of units of graph space between bars.  Should be 
-		/// less than <see cref="BarSeries.BarEvery"/>
+		/// less than <see cref="BarEvery"/>
 		/// </summary>
 		public float Spacing { get; }
 
@@ -119,7 +119,7 @@ namespace Terminal.Gui {
 			}
 
 			for (int i = 0; i < values.Length; i++) {
-				subSeries [i].Bars.Add (new BarSeries.Bar (label,
+				subSeries [i].Bars.Add (new Bar (label,
 					new GraphCellToRender (fill), values [i]));
 			}
 		}
@@ -142,7 +142,7 @@ namespace Terminal.Gui {
 	/// <summary>
 	/// Series of bars positioned at regular intervals
 	/// </summary>
-	public class BarSeries : ISeries {
+	public partial class BarSeries : ISeries {
 
 		/// <summary>
 		/// Ordered collection of graph bars to position along axis
@@ -279,43 +279,6 @@ namespace Terminal.Gui {
 			graph.DrawLine (start, end, adjusted.Rune);
 
 			graph.SetDriverColorToGraphColor ();
-		}
-
-		/// <summary>
-		/// A single bar in a <see cref="BarSeries"/>
-		/// </summary>
-		public class Bar {
-
-			/// <summary>
-			/// Optional text that describes the bar.  This will be rendered on the corresponding
-			/// <see cref="Axis"/> unless <see cref="DrawLabels"/> is false
-			/// </summary>
-			public string Text { get; set; }
-
-			/// <summary>
-			/// The color and character that will be rendered in the console
-			/// when the bar extends over it
-			/// </summary>
-			public GraphCellToRender Fill { get; set; }
-
-			/// <summary>
-			/// The value in graph space X/Y (depending on <see cref="Orientation"/>) to which the bar extends.
-			/// </summary>
-			public float Value { get; }
-
-			/// <summary>
-			/// Creates a new instance of a single bar rendered in the given <paramref name="fill"/> that extends
-			/// out <paramref name="value"/> graph space units in the default <see cref="Orientation"/>
-			/// </summary>
-			/// <param name="text"></param>
-			/// <param name="fill"></param>
-			/// <param name="value"></param>
-			public Bar (string text, GraphCellToRender fill, float value)
-			{
-				Text = text;
-				Fill = fill;
-				Value = value;
-			}
 		}
 	}
 }
