+++ Terminal.Gui/Views/TableView/ListTableSource.cs
@@ -9,7 +9,7 @@ namespace Terminal.Gui {
 	/// a <see cref="System.Collections.IList"/>.  This class is
 	/// mutable: changes are permitted to the wrapped <see cref="IList"/>.
 	/// </summary>
-	public class ListTableSource : ITableSource {
+	public partial class ListTableSource : ITableSource {
 		/// <summary>
 		/// The list this source wraps.
 		/// </summary>
@@ -173,24 +173,6 @@ namespace Terminal.Gui {
 			}
 
 			return (cols > 1) ? cols : 1;
-		}
-
-		/// <summary>
-		/// Defines rendering options that affect how the view is displayed.
-		/// </summary>
-		public class ListColumnStyle {
-
-			/// <summary>
-			/// Gets or sets an Orientation enum indicating whether to populate data down each column
-			/// rather than across each row.  Defaults to <see cref="Orientation.Horizontal"/>.
-			/// </summary>
-			public Orientation Orientation { get; set; } = Orientation.Horizontal;
-
-			/// <summary>
-			/// Gets or sets a flag indicating whether to scroll in the same direction as <see cref="ListTableSource.ListColumnStyle.Orientation"/>.
-			/// Defaults to <see langword="false"/>.
-			/// </summary>
-			public bool ScrollParallel { get; set; } = false;
 		}
 	}
 }
