+++ UnitTests/Dialogs/WizardTests.cs
@@ -159,7 +159,7 @@ namespace Terminal.Gui.DialogTests {
 			var bottomRow = $"{CM.Glyphs.LLCornerDbl}{new string (CM.Glyphs.HLineDbl.ToString () [0], width - 2)}{CM.Glyphs.LRCornerDbl}";
 
 			var wizard = new Wizard () { Title = title, Width = width, Height = height };
-			wizard.AddStep (new Wizard.WizardStep () { Title = stepTitle });
+			wizard.AddStep (new WizardStep () { Title = stepTitle });
 			//wizard.LayoutSubviews ();
 			var firstIteration = false;
 			var runstate = Application.Begin (wizard);
@@ -229,7 +229,7 @@ namespace Terminal.Gui.DialogTests {
 			var bottomRow = $"{CM.Glyphs.LLCornerDbl}{new string (CM.Glyphs.HLineDbl.ToString () [0], width - 2)}{CM.Glyphs.LRCornerDbl}";
 
 			var wizard = new Wizard () { Title = title, Width = width, Height = height };
-			wizard.AddStep (new Wizard.WizardStep () { Title = "ABCD" });
+			wizard.AddStep (new WizardStep () { Title = "ABCD" });
 
 			Application.End (Application.Begin (wizard));
 			TestHelpers.AssertDriverContentsWithFrameAre ($"{topRow}\n{separatorRow}\n{buttonRow}\n{bottomRow}", output);
@@ -243,7 +243,7 @@ namespace Terminal.Gui.DialogTests {
 			// If no steps should be null
 			Assert.Null (wizard.GetPreviousStep ());
 
-			var step1 = new Wizard.WizardStep () { Title = "step1" };
+			var step1 = new WizardStep () { Title = "step1" };
 			wizard.AddStep (step1);
 
 			// If no current step, should be last step
@@ -258,7 +258,7 @@ namespace Terminal.Gui.DialogTests {
 			Assert.Null (wizard.GetPreviousStep ());
 
 			// If two steps and at 2 and step 1 is `Enabled = true`should be step1
-			var step2 = new Wizard.WizardStep () { Title = "step2" };
+			var step2 = new WizardStep () { Title = "step2" };
 			wizard.AddStep (step2);
 			wizard.CurrentStep = step2;
 			step1.Enabled = true;
@@ -272,7 +272,7 @@ namespace Terminal.Gui.DialogTests {
 			//   At step 1 should be null
 			//   At step 2 should be step 1
 			//   At step 3 should be step 2
-			var step3 = new Wizard.WizardStep () { Title = "step3" };
+			var step3 = new WizardStep () { Title = "step3" };
 			wizard.AddStep (step3);
 			step1.Enabled = true;
 			wizard.CurrentStep = step1;
@@ -334,7 +334,7 @@ namespace Terminal.Gui.DialogTests {
 			// If no steps should be null
 			Assert.Null (wizard.GetNextStep ());
 
-			var step1 = new Wizard.WizardStep () { Title = "step1" };
+			var step1 = new WizardStep () { Title = "step1" };
 			wizard.AddStep (step1);
 
 			// If no current step, should be first step
@@ -349,7 +349,7 @@ namespace Terminal.Gui.DialogTests {
 			Assert.Null (wizard.GetNextStep ());
 
 			// If two steps and at 1 and step 2 is `Enabled = true`should be step 2
-			var step2 = new Wizard.WizardStep () { Title = "step2" };
+			var step2 = new WizardStep () { Title = "step2" };
 			wizard.AddStep (step2);
 			Assert.Equal (step2.Title, wizard.GetNextStep ().Title);
 
@@ -363,7 +363,7 @@ namespace Terminal.Gui.DialogTests {
 			//   At step 1 should be step 2
 			//   At step 2 should be step 3
 			//   At step 3 should be null
-			var step3 = new Wizard.WizardStep () { Title = "step3" };
+			var step3 = new WizardStep () { Title = "step3" };
 			wizard.AddStep (step3);
 			step1.Enabled = true;
 			wizard.CurrentStep = step1;
@@ -457,15 +457,15 @@ namespace Terminal.Gui.DialogTests {
 
 			Assert.Null (wizard.GetFirstStep ());
 
-			var step1 = new Wizard.WizardStep () { Title = "step1" };
+			var step1 = new WizardStep () { Title = "step1" };
 			wizard.AddStep (step1);
 			Assert.Equal (step1.Title, wizard.GetFirstStep ().Title);
 
-			var step2 = new Wizard.WizardStep () { Title = "step2" };
+			var step2 = new WizardStep () { Title = "step2" };
 			wizard.AddStep (step2);
 			Assert.Equal (step1.Title, wizard.GetFirstStep ().Title);
 
-			var step3 = new Wizard.WizardStep () { Title = "step3" };
+			var step3 = new WizardStep () { Title = "step3" };
 			wizard.AddStep (step3);
 			Assert.Equal (step1.Title, wizard.GetFirstStep ().Title);
 
@@ -487,15 +487,15 @@ namespace Terminal.Gui.DialogTests {
 
 			Assert.Null (wizard.GetLastStep ());
 
-			var step1 = new Wizard.WizardStep () { Title = "step1" };
+			var step1 = new WizardStep () { Title = "step1" };
 			wizard.AddStep (step1);
 			Assert.Equal (step1.Title, wizard.GetLastStep ().Title);
 
-			var step2 = new Wizard.WizardStep () { Title = "step2" };
+			var step2 = new WizardStep () { Title = "step2" };
 			wizard.AddStep (step2);
 			Assert.Equal (step2.Title, wizard.GetLastStep ().Title);
 
-			var step3 = new Wizard.WizardStep () { Title = "step3" };
+			var step3 = new WizardStep () { Title = "step3" };
 			wizard.AddStep (step3);
 			Assert.Equal (step3.Title, wizard.GetLastStep ().Title);
 
@@ -515,7 +515,7 @@ namespace Terminal.Gui.DialogTests {
 		{
 			// https://github.com/gui-cs/Terminal.Gui/issues/1833
 			var wizard = new Wizard ();
-			var step1 = new Wizard.WizardStep () { Title = "step1" };
+			var step1 = new WizardStep () { Title = "step1" };
 			wizard.AddStep (step1);
 
 			var finishedFired = false;
@@ -543,9 +543,9 @@ namespace Terminal.Gui.DialogTests {
 			// Same test, but with two steps
 			wizard = new Wizard ();
 			firstIteration = false;
-			step1 = new Wizard.WizardStep () { Title = "step1" };
+			step1 = new WizardStep () { Title = "step1" };
 			wizard.AddStep (step1);
-			var step2 = new Wizard.WizardStep () { Title = "step2" };
+			var step2 = new WizardStep () { Title = "step2" };
 			wizard.AddStep (step2);
 
 			finishedFired = false;
@@ -580,9 +580,9 @@ namespace Terminal.Gui.DialogTests {
 			// Same test, but with two steps but the 1st one disabled
 			wizard = new Wizard ();
 			firstIteration = false;
-			step1 = new Wizard.WizardStep () { Title = "step1" };
+			step1 = new WizardStep () { Title = "step1" };
 			wizard.AddStep (step1);
-			step2 = new Wizard.WizardStep () { Title = "step2" };
+			step2 = new WizardStep () { Title = "step2" };
 			wizard.AddStep (step2);
 			step1.Enabled = false;
